% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{answer_2subset}
\alias{answer_2subset}
\title{answer_2subset}
\format{\preformatted{Classes <U+0091>data.table<U+0092> and 'data.frame':	246 obs. of  4 variables:
 $ school_year     : chr  "2011" "2011" "2011" "2011" ...
 $ fake_student_id : chr  "109121" "109121" "109121" "109121" ...
 $ test_grade      : chr  "8" "8" "8" "8" ...
 $ test_scale_score: chr  "895" "848" "870" "0" ...
 - attr(*, ".internal.selfref")=<externalptr> 
}}
\usage{
answer_2subset
}
\description{
A dataset that can be used to check your answers to the do it yourself section of 2_subset.
}
\keyword{datasets}

