% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{answer_3merge}
\alias{answer_3merge}
\title{answer_3merge}
\format{\preformatted{Classes <U+0091>data.table<U+0092> and 'data.frame':	92 obs. of  24 variables:
 $ fake_student_id       : chr  "105989" "105989" "105989" "105989" ...
 $ school_year           : chr  "2011" "2011" "2012" "2012" ...
 $ fake_school_id        : chr  "89" "89" "133" "133" ...
 $ test_name             : chr  "State Test" "State Test" "State Test" "State Test" ...
 $ test_period           : chr  "3" "3" "3" "3" ...
 $ opportunity_number    : chr  "1" "1" "4" "3" ...
 $ test_grade            : chr  "6" "6" "7" "7" ...
 $ test_content_area     : chr  "ELA" "MAT" "ELA" "ELA" ...
 $ test_start_date       : chr  "5/6/2011" "5/16/2011" "4/27/2012" "3/16/2012" ...
 $ test_end_date         : chr  "5/6/2011" "5/16/2011" "4/27/2012" "3/16/2012" ...
 $ test_sem              : chr  "23.018124" "19.237476" "25.794645" "23.991076" ...
 $ test_scale_score      : chr  "693" "732" "696" "696" ...
 $ test_performance_level: chr  "1" "2" "1" "1" ...
 $ test_raw_score        : chr  "22" "25" "18" "22" ...
 $ gender                : chr  "F" "F" "F" "F" ...
 $ race                  : chr  "2" "2" "2" "2" ...
 $ race_descr            : chr  "Black" "Black" "Black" "Black" ...
 $ migrant_ind           : chr  "N" "N" "N" "N" ...
 $ homeless_ind          : chr  "N" "N" "N" "N" ...
 $ sped_ind              : chr  "Y" "Y" "Y" "Y" ...
 $ sped_code             : chr  "300" "300" "300" "300" ...
 $ sped_descr            : chr  "Learning Disability" "Learning Disability" "Learning Disability" "Learning Disability" ...
 $ low_income_ind        : chr  "Y" "Y" "Y" "Y" ...
 $ ell_status            : chr  "N" "N" "N" "N" ...
 - attr(*, "sorted")= chr  "fake_student_id" "school_year"
 - attr(*, ".internal.selfref")=<externalptr> 
}}
\usage{
answer_3merge
}
\description{
A dataset that can be used to check your answers to the do it yourself section of 3_merge.
}
\keyword{datasets}

