% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{answer_4recode}
\alias{answer_4recode}
\title{answer_4recode}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 11 rows and 14 columns.
}
\usage{
answer_4recode
}
\description{
A dataset that can be used to check your answers to the do it yourself section of 4_recode.
}
\keyword{datasets}
