% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{answer_5operations}
\alias{answer_5operations}
\title{answer_5operations}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 130 rows and 15 columns.
}
\usage{
answer_5operations
}
\description{
A dataset that can be used to check your answers to the do it yourself section of 5_operations.
}
\keyword{datasets}
