% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{answer_6bygroup}
\alias{answer_6bygroup}
\title{answer_6bygroup}
\format{\preformatted{Classes <U+0091>data.table<U+0092> and 'data.frame':	4 obs. of  3 variables:
 $ school_year: chr  "2013" "2012" "2011" "2014"
 $ avg_perform: num  2.84 2.79 2.38 2.53
 $ count      : int  108 146 188 58
 - attr(*, ".internal.selfref")=<externalptr> 
}}
\usage{
answer_6bygroup
}
\description{
A dataset that can be used to check your answers to the do it yourself section of 6_by_group.
}
\keyword{datasets}

