% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{answer_7transpose}
\alias{answer_7transpose}
\title{answer_7transpose}
\format{\preformatted{Classes <U+0091>data.table<U+0092> and 'data.frame':	500 obs. of  7 variables:
 $ fake_student_id       : int  95232 95232 106894 106894 96252 96252 96252 96252 96252 96252 ...
 $ test_name             : chr  "State Test" "State Test" "State Test" "State Test" ...
 $ opportunity_number    : int  1 1 1 1 1 1 1 1 1 1 ...
 $ test_period           : int  3 3 3 3 3 3 1 1 2 2 ...
 $ test_content_area     : chr  "ELA" "MAT" "ELA" "MAT" ...
 $ year_test_grade       : Factor w/ 30 levels "2011_2","2011_3",..: 1 1 1 1 2 2 2 2 2 2 ...
 $ test_performance_level: int  3 4 4 4 0 0 1 1 2 2 ...
 - attr(*, ".internal.selfref")=<externalptr> 
}}
\usage{
answer_7transpose
}
\description{
A dataset that can be used to check your answers to the do it yourself section of 7_transpose.
}
\keyword{datasets}

