% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{answer_9loop}
\alias{answer_9loop}
\title{answer_9loop}
\format{\preformatted{ num 9885
}}
\usage{
answer_9loop
}
\description{
A variable that can be used to check your answers to the do it yourself section of 9_loop.
}
\keyword{datasets}

