% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_ggplot2}
\alias{data_ggplot2}
\title{data_ggplot2}
\format{\preformatted{'data.frame':	40559 obs. of  8 variables:
 $ student_id           : int  99002346 99033513 99007189 99015450 99036312 99013203 99000796 99042711 99031307 99021655 ...
 $ school_id            : int  50 122 337 45 41 237 47 42 46 338 ...
 $ tested_grade         : int  7 5 6 6 6 5 8 8 8 8 ...
 $ math_scale_score_2013: int  2006 1416 1799 1575 1461 1341 1562 1448 1551 1516 ...
 $ read_scale_score_2013: int  1643 1422 1713 1505 1556 1251 1573 1460 1438 1448 ...
 $ d_gender_m           : int  1 1 1 0 1 1 0 1 1 0 ...
 $ d_econdisadv_y       : int  0 1 1 1 1 1 1 1 1 1 ...
 $ d_lep_y              : int  0 0 0 0 0 0 0 0 0 0 ...
}}
\usage{
data_ggplot2
}
\description{
Demographic and test data of students. One student per row of data.
Dataset has 40,559 students and 8 variables.
}
\keyword{datasets}

