% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{DelawareDataPackage}
\alias{DelawareDataPackage}
\title{Data package for training code.}
\description{
This package contains data that can be used with the training code.
The names of the data included are bellow.
}
\seealso{
\code{\link{answer_2subset}}

\code{\link{answer_3merge}}

\code{\link{answer_4recode}}

\code{\link{answer_5operations}}

\code{\link{answer_6bygroup}}

\code{\link{answer_7transpose}}

\code{\link{answer_9loop}}

\code{\link{data_ggplot2}}

\code{\link{dems_single_row_student}}

\code{\link{dems_single_row_student_wgrade}}

\code{\link{dems_single_row_year}}

\code{\link{test_multi_row}}

\code{\link{test_single_row}}
}
