% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dems_single_row_student}
\alias{dems_single_row_student}
\title{dems_single_row_student.}
\format{\preformatted{Classes <U+0091>data.table<U+0092> and 'data.frame':	11 obs. of  11 variables:
 $ fake_student_id: chr  "114811" "86205" "105989" "95232" ...
 $ gender         : chr  "F" "M" "F" "M" ...
 $ race           : chr  "5" "2" "2" "5" ...
 $ race_descr     : chr  "White" "Black" "Black" "White" ...
 $ migrant_ind    : chr  "N" "N" "N" "N" ...
 $ homeless_ind   : chr  "N" "N" "N" "N" ...
 $ sped_ind       : chr  "N" "N" "Y" "N" ...
 $ sped_code      : chr  "0" "0" "300" "0" ...
 $ sped_descr     : chr  "Regular-Ed" "Regular-Ed" "Learning Disability" "Regular-Ed" ...
 $ low_income_ind : chr  "" "" "" "Y" ...
 $ ell_status     : chr  "N" "N" "N" "N" ...
 - attr(*, ".internal.selfref")=<externalptr> 
}}
\usage{
dems_single_row_student
}
\description{
Demographic data of students. One student per row of data.
Short dataset with only 11 students and 11 variables.
}
\keyword{datasets}

