% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dems_single_row_student_wgrade}
\alias{dems_single_row_student_wgrade}
\title{dems_single_row_student_wgrade.}
\format{\preformatted{'data.frame':	11 obs. of  12 variables:
 $ fake_student_id: int  114811 86205 105989 95232 91134 115180 100065 96592 106894 109944 ...
 $ gender         : Factor w/ 2 levels "F","M": 1 2 1 2 2 2 1 2 2 2 ...
 $ race           : int  5 2 2 5 2 5 5 6 5 5 ...
 $ race_descr     : Factor w/ 3 levels "Asian","Black",..: 3 2 2 3 2 3 3 1 3 3 ...
 $ migrant_ind    : Factor w/ 1 level "N": 1 1 1 1 1 1 1 1 1 1 ...
 $ homeless_ind   : Factor w/ 1 level "N": 1 1 1 1 1 1 1 1 1 1 ...
 $ sped_ind       : Factor w/ 2 levels "N","Y": 1 1 2 1 1 2 1 1 1 1 ...
 $ sped_code      : int  0 0 300 0 0 300 0 0 0 0 ...
 $ sped_descr     : Factor w/ 2 levels "Learning Disability",..: 2 2 1 2 2 1 2 2 2 2 ...
 $ low_income_ind : Factor w/ 3 levels "","N","Y": 1 1 1 3 2 2 2 2 1 2 ...
 $ ell_status     : Factor w/ 1 level "N": 1 1 1 1 1 1 1 1 1 1 ...
 $ grade          : int  7 6 4 9 3 6 5 3 11 9 ...
}}
\usage{
dems_single_row_student_wgrade
}
\description{
Demographic data of students. One student per row of data.
Short dataset with only 11 students and 12 variables.
}
\keyword{datasets}

