% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dems_single_row_year}
\alias{dems_single_row_year}
\title{dems_single_row_year}
\format{\preformatted{'data.frame':	74 obs. of  12 variables:
 $ fake_student_id: int  114811 114811 114811 114811 86205 86205 86205 86205 89244 89244 ...
 $ school_year    : int  2012 2011 2013 2014 2013 2012 2011 2014 2011 2012 ...
 $ gender         : Factor w/ 2 levels "F","M": 1 1 1 1 2 2 2 2 1 1 ...
 $ race           : int  5 5 5 5 2 2 2 2 5 5 ...
 $ race_descr     : Factor w/ 3 levels "Asian","Black",..: 3 3 3 3 2 2 2 2 3 3 ...
 $ migrant_ind    : Factor w/ 1 level "N": 1 1 1 1 1 1 1 1 1 1 ...
 $ homeless_ind   : Factor w/ 1 level "N": 1 1 1 1 1 1 1 1 1 1 ...
 $ sped_ind       : Factor w/ 2 levels "N","Y": 1 1 1 1 1 1 1 1 1 1 ...
 $ sped_code      : int  0 0 0 0 0 0 0 0 0 0 ...
 $ sped_descr     : Factor w/ 3 levels "Learning Disability",..: 3 3 3 3 3 3 3 3 3 3 ...
 $ low_income_ind : Factor w/ 2 levels "N","Y": 2 2 2 2 1 1 1 1 2 2 ...
 $ ell_status     : Factor w/ 2 levels "N","Y": 1 1 1 1 1 1 1 1 1 1 ...
}}
\usage{
dems_single_row_year
}
\description{
Demographic data of students. One student per row of data.
Dataset has 74 students and 12 variables.
}
\keyword{datasets}

