% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_multi_row}
\alias{test_multi_row}
\title{test_multi_row}
\format{\preformatted{Classes <U+0091>data.table<U+0092> and 'data.frame':	500 obs. of  14 variables:
 $ school_year           : chr  "2013" "2012" "2012" "2011" ...
 $ fake_student_id       : chr  "100140" "100140" "89244" "109121" ...
 $ fake_school_id        : chr  "103" "103" "70" "135" ...
 $ test_name             : chr  "State Test" "State Test" "State Test" "State Test" ...
 $ test_period           : chr  "3" "3" "3" "1" ...
 $ opportunity_number    : chr  "1" "1" "1" "1" ...
 $ test_grade            : chr  "11" "10" "10" "8" ...
 $ test_content_area     : chr  "Alt" "Alt" "Alt" "ELA" ...
 $ test_start_date       : chr  "5/7/2013" "5/10/2012" "5/21/2012" "10/19/2010" ...
 $ test_end_date         : chr  "5/7/2013" "5/10/2012" "5/21/2012" "10/19/2010" ...
 $ test_sem              : chr  "0" "10.710167" "10.712411" "29.84989" ...
 $ test_scale_score      : chr  "0" "419" "361" "895" ...
 $ test_performance_level: chr  "1" "3" "1" "4" ...
 $ test_raw_score        : chr  NA NA NA "42" ...
 - attr(*, ".internal.selfref")=<externalptr> 
}}
\usage{
test_multi_row
}
\description{
Demographic data of students. One student per row of data.
Dataset has 500 students and 14 variables.
}
\keyword{datasets}

