% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{answer_2subset}
\alias{answer_2subset}
\title{answer_2subset}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 246 rows and 4 columns.
}
\usage{
answer_2subset
}
\description{
A dataset that can be used to check your answers to the do it yourself section of 2_subset.
}
\keyword{datasets}
