% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{answer_3merge}
\alias{answer_3merge}
\title{answer_3merge}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 92 rows and 24 columns.
}
\usage{
answer_3merge
}
\description{
A dataset that can be used to check your answers to the do it yourself section of 3_merge.
}
\keyword{datasets}
