% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{answer_4recode}
\alias{answer_4recode}
\title{answer_4recode}
\format{\preformatted{Classes <U+0091>data.table<U+0092> and 'data.frame':	11 obs. of  14 variables:
 $ fake_student_id: int  114811 86205 105989 95232 91134 115180 100065 96592 106894 109944 ...
 $ gender         : chr  "F" "M" "F" "M" ...
 $ race           : int  5 2 2 5 2 5 5 6 5 5 ...
 $ race_descr     : chr  "White" "Black" "Black" "White" ...
 $ migrant_ind    : chr  "N" "N" "N" "N" ...
 $ homeless_ind   : chr  "N" "N" "N" "N" ...
 $ sped_ind       : chr  "N" "N" "Y" "N" ...
 $ sped_code      : int  0 0 300 0 0 300 0 0 0 0 ...
 $ sped_descr     : chr  "Regular-Ed" "Regular-Ed" "Learning Disability" "Regular-Ed" ...
 $ low_income_ind : chr  "" "" "" "Y" ...
 $ ell_status     : chr  "N" "N" "N" "N" ...
 $ grade          : int  7 6 4 9 3 6 5 3 11 9 ...
 $ bi_gender      : chr  "1" "0" "1" "0" ...
 $ lower_upper    : chr  "upper" "lower" "lower" "upper" ...
 - attr(*, ".internal.selfref")=<externalptr> 
 - attr(*, "index")= atomic  
  ..- attr(*, "__gender")= int  1 3 7 2 4 5 6 8 9 10 ...
  ..- attr(*, "__grade")= int  5 8 3 7 11 2 6 1 4 10 ...
}}
\usage{
answer_4recode
}
\description{
A dataset that can be used to check your answers to the do it yourself section of 4_recode.
}
\keyword{datasets}

