% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{answer_5operations}
\alias{answer_5operations}
\title{answer_5operations}
\format{\preformatted{Classes <U+0091>data.table<U+0092> and 'data.frame':	130 obs. of  15 variables:
 $ school_year           : chr  "2011" "2011" "2011" "2011" ...
 $ fake_student_id       : chr  "109121" "109944" "97475" "115025" ...
 $ fake_school_id        : chr  "135" "81" "81" "81" ...
 $ test_name             : chr  "State Test" "State Test" "State Test" "State Test" ...
 $ test_period           : chr  "1" "1" "1" "1" ...
 $ opportunity_number    : chr  "1" "1" "1" "1" ...
 $ test_grade            : chr  "8" "8" "8" "8" ...
 $ test_content_area     : chr  "ELA" "ELA" "ELA" "ELA" ...
 $ test_start_date       : chr  "10/19/2010" "10/25/2010" "10/19/2010" "10/20/2010" ...
 $ test_end_date         : chr  "10/19/2010" "10/25/2010" "10/19/2010" "10/20/2010" ...
 $ test_sem              : chr  "29.84989" "25.430241" "22.920309" "28.375031" ...
 $ test_scale_score      : num  895 853 743 884 912 936 928 831 782 844 ...
 $ test_performance_level: num  4 4 1 4 4 4 4 3 2 4 ...
 $ test_raw_score        : num  42 38 25 41 42 44 45 36 30 34 ...
 $ arbitrary_math        : num  -895 -853 -743 -884 -912 ...
 - attr(*, ".internal.selfref")=<externalptr> 
}}
\usage{
answer_5operations
}
\description{
A dataset that can be used to check your answers to the do it yourself section of 5_operations.
}
\keyword{datasets}

