% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{answer_6bygroup}
\alias{answer_6bygroup}
\title{answer_6bygroup}
\format{\preformatted{Classes <U+0091>data.table<U+0092> and 'data.frame':	41 obs. of  5 variables:
 $ school_year           : chr  "2011" "2011" "2011" "2011" ...
 $ avg_perform           : num  2.38 2.38 2.38 2.38 2.38 ...
 $ test_period           : chr  "1" "3" "2" "3" ...
 $ test_performance_level: chr  "4" "4" "4" "0" ...
 $ count                 : int  1 1 1 1 1 1 1 1 1 1 ...
 - attr(*, "sorted")= chr "school_year"
 - attr(*, ".internal.selfref")=<externalptr> 
}}
\usage{
answer_6bygroup
}
\description{
A dataset that can be used to check your answers to the do it yourself section of 6_by_group.
}
\keyword{datasets}

