% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{answer_6bygroup}
\alias{answer_6bygroup}
\title{answer_6bygroup}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 4 rows and 3 columns.
}
\usage{
answer_6bygroup
}
\description{
A dataset that can be used to check your answers to the do it yourself section of 6_by_group.
}
\keyword{datasets}
