% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{answer_7transpose}
\alias{answer_7transpose}
\title{answer_7transpose}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 500 rows and 7 columns.
}
\usage{
answer_7transpose
}
\description{
A dataset that can be used to check your answers to the do it yourself section of 7_transpose.
}
\keyword{datasets}
