% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_ggplot2}
\alias{data_ggplot2}
\title{data_ggplot2}
\format{\preformatted{Classes <U+0091>data.table<U+0092> and 'data.frame':	40559 obs. of  8 variables:
 $ student_id           : chr  "99002346" "99033513" "99007189" "99015450" ...
 $ school_id            : chr  "50" "122" "337" "45" ...
 $ tested_grade         : chr  "7" "5" "6" "6" ...
 $ math_scale_score_2013: chr  "2006" "1416" "1799" "1575" ...
 $ read_scale_score_2013: chr  "1643" "1422" "1713" "1505" ...
 $ d_gender_m           : chr  "1" "1" "1" "0" ...
 $ d_econdisadv_y       : chr  "0" "1" "1" "1" ...
 $ d_lep_y              : chr  "0" "0" "0" "0" ...
 - attr(*, ".internal.selfref")=<externalptr> 
}}
\usage{
data_ggplot2
}
\description{
Demographic and test data of students. One student per row of data.
Dataset has 40,559 students and 8 variables.
}
\keyword{datasets}

