% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_ggplot2}
\alias{data_ggplot2}
\title{data_ggplot2}
\format{A data frame with eight variables:  \code{student_id},
\code{school_id}, \code{tested_grade}, \code{math_scale_score_2013},
\code{read_scale_score_2013}, \code{d_gender_m}, \code{d_econdisadv_y},
and \code{d_lep_y}.}
\usage{
data_ggplot2
}
\description{
Demographic and test data of students. One student per row of data.
Dataset has 40,559 students and 8 variables.
}
\keyword{datasets}

