% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_ggplot2}
\alias{data_ggplot2}
\title{data_ggplot2}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 40559 rows and 8 columns.
}
\usage{
data_ggplot2
}
\description{
Demographic and test data of students. One student per row of data.
Dataset has 40,559 students and 8 variables.
}
\keyword{datasets}
