% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dems_single_row_student}
\alias{dems_single_row_student}
\title{dems_single_row_student.}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 11 rows and 11 columns.
}
\usage{
dems_single_row_student
}
\description{
Demographic data of students. One student per row of data.
Short dataset with only 11 students and 11 variables.
}
\keyword{datasets}
