% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dems_single_row_year}
\alias{dems_single_row_year}
\title{dems_single_row_year}
\format{A data frame with twelve variables:  \code{fake_student_id},
 \code{school_year}, \code{gender}, \code{race}, \code{race_descr}, \code{migrant_ind},
 \code{homeless_ind}, \code{sped_ind}, \code{sped_code}, \code{sped_descr},
 \code{low_income_ind}, and \code{ell_status}.}
\usage{
dems_single_row_year
}
\description{
Demographic data of students. One student per row of data.
Dataset has 74 students and 12 variables.
}
\keyword{datasets}

