% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dems_single_row_year}
\alias{dems_single_row_year}
\title{dems_single_row_year}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 74 rows and 12 columns.
}
\usage{
dems_single_row_year
}
\description{
Demographic data of students. One student per row of data.
Dataset has 74 students and 12 variables.
}
\keyword{datasets}
