% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dems_single_row_year}
\alias{dems_single_row_year}
\title{dems_single_row_year}
\format{\preformatted{Classes <U+0091>data.table<U+0092> and 'data.frame':	74 obs. of  12 variables:
 $ fake_student_id: chr  "114811" "114811" "114811" "114811" ...
 $ school_year    : chr  "2012" "2011" "2013" "2014" ...
 $ gender         : chr  "F" "F" "F" "F" ...
 $ race           : chr  "5" "5" "5" "5" ...
 $ race_descr     : chr  "White" "White" "White" "White" ...
 $ migrant_ind    : chr  "N" "N" "N" "N" ...
 $ homeless_ind   : chr  "N" "N" "N" "N" ...
 $ sped_ind       : chr  "N" "N" "N" "N" ...
 $ sped_code      : chr  "0" "0" "0" "0" ...
 $ sped_descr     : chr  "Regular-Ed" "Regular-Ed" "Regular-Ed" "Regular-Ed" ...
 $ low_income_ind : chr  "Y" "Y" "Y" "Y" ...
 $ ell_status     : chr  "N" "N" "N" "N" ...
 - attr(*, ".internal.selfref")=<externalptr> 
}}
\usage{
dems_single_row_year
}
\description{
Demographic data of students. One student per row of data.
Dataset has 74 students and 12 variables.
}
\keyword{datasets}

