% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_multi_row}
\alias{test_multi_row}
\title{test_multi_row}
\format{\preformatted{'data.frame':	500 obs. of  14 variables:
 $ school_year           : int  2013 2012 2012 2011 2011 2011 2011 2011 2011 2011 ...
 $ fake_student_id       : int  100140 100140 89244 109121 109121 109121 109121 97475 109944 97475 ...
 $ fake_school_id        : int  103 103 70 135 135 135 135 81 81 81 ...
 $ test_name             : Factor w/ 2 levels "Alternate Test",..: 2 2 2 2 2 2 1 2 2 1 ...
 $ test_period           : int  3 3 3 1 3 2 3 2 1 3 ...
 $ opportunity_number    : int  1 1 1 1 1 1 1 1 1 1 ...
 $ test_grade            : int  11 10 10 8 8 8 8 8 8 8 ...
 $ test_content_area     : Factor w/ 7 levels "Alt","Alt2","ELA",..: 1 1 1 3 3 3 3 3 3 3 ...
 $ test_start_date       : Factor w/ 241 levels "1/10/2011","1/10/2012",..: 204 148 178 43 163 79 198 1 49 170 ...
 $ test_end_date         : Factor w/ 241 levels "1/10/2011","1/10/2012",..: 204 148 178 43 163 79 198 1 49 170 ...
 $ test_sem              : num  0 10.7 10.7 29.8 26.5 ...
 $ test_scale_score      : int  0 419 361 895 848 870 0 709 853 0 ...
 $ test_performance_level: int  1 3 1 4 4 4 0 1 4 0 ...
 $ test_raw_score        : int  NA NA NA 42 39 42 0 17 38 0 ...
}}
\usage{
test_multi_row
}
\description{
Demographic data of students. One student per row of data.
Dataset has 500 students and 14 variables.
}
\keyword{datasets}

