% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_multi_row}
\alias{test_multi_row}
\title{test_multi_row}
\format{A data frame with fourteen variables:  \code{school_year},
\code{fake_student_id}, \code{fake_school_id}, \code{test_name},
\code{test_period}, \code{opportunity_number}, \code{test_grade},
\code{test_content_area}, \code{test_start_date}, \code{test_end_date},
\code{test_sem}, \code{test_scale_score}, \code{test_performance_level}, and \code{test_raw_score}.}
\usage{
test_multi_row
}
\description{
Demographic data of students. One student per row of data.
Dataset has 500 students and 14 variables.
}
\keyword{datasets}

