% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_single_row}
\alias{test_single_row}
\title{test_single_row}
\format{\preformatted{Classes <U+0091>data.table<U+0092> and 'data.frame':	8 obs. of  14 variables:
 $ school_year           : chr  "2014" "2014" "2014" "2014" ...
 $ fake_student_id       : chr  "100065" "95232" "106894" "107885" ...
 $ fake_school_id        : chr  "80" "99" "87" "91" ...
 $ test_name             : chr  "State Test" "State Test" "State Test" "State Test" ...
 $ test_period           : chr  "3" "3" "3" "3" ...
 $ opportunity_number    : chr  "3" "3" "3" "3" ...
 $ test_grade            : chr  "10" "5" "5" "9" ...
 $ test_content_area     : chr  "MAT" "MAT" "MAT" "MAT" ...
 $ test_start_date       : chr  "5/21/2014 8:02" "5/7/2014 13:43" "5/12/2014 10:25" "5/21/2014 7:30" ...
 $ test_end_date         : chr  "" "" "" "" ...
 $ test_sem              : chr  "19.8297" "20.2068" "25.9666" "19.8323" ...
 $ test_scale_score      : chr  "900" "769" "850" "793" ...
 $ test_performance_level: chr  "4" "3" "4" "2" ...
 $ test_raw_score        : chr  "32" "29" "41" "22" ...
 - attr(*, ".internal.selfref")=<externalptr> 
}}
\usage{
test_single_row
}
\description{
Demographic data of students. One student per row of data.
Dataset has 8 students and 14 variables.
}
\keyword{datasets}

