% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_single_row}
\alias{test_single_row}
\title{test_single_row}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 8 rows and 14 columns.
}
\usage{
test_single_row
}
\description{
Demographic data of students. One student per row of data.
Dataset has 8 students and 14 variables.
}
\keyword{datasets}
